requirements.txt
anthropic
google-genai
google-auth-oauthlib
google-api-python-client
python-dotenv
flask
